+++ src/lightwin/config/key_val_conf_spec.py
@@ -99,7 +99,10 @@ class KeyValConfSpec:
         """Check that the value has the proper typing."""
         if isinstance(toml_value, self.types):
             return True
-        logging.error(f"{toml_value = } type not in {self.types = }")
+        logging.warning(
+            f"Type error in {self.key}. {toml_value = } type not in "
+            f"{self.types = }"
+        )
         return False
 
     def is_valid_value(self, toml_value: Any, **kwargs) -> bool:
@@ -108,7 +111,9 @@ class KeyValConfSpec:
             return True
         if toml_value in self.allowed_values:
             return True
-        logging.error(f"{toml_value = } is not in {self.allowed_values = }")
+        logging.error(
+            f"{self.key}: {toml_value = } is not in {self.allowed_values = }"
+        )
         return False
 
     def path_exists(